FIX INTRO SCENE + HARDCODE

+++ 05-SceneManager/HUD.cpp
@@ -12,7 +12,7 @@ void HUD::Draw()
 		if (!MarioIsFlying) {
 			for (int i = 0; i < speedStack; i++)
 			{
-				float startX = 62 + i * 8;
+				float startX = 62.0f + i * 8;
 				if (i == 6)
 					sprites->Get(ID_SPEED_STACK_P)->Draw(Camera::GetInstance()->cam_x + startX + 4, Camera::GetInstance()->cam_y + 210);
 				else
@@ -24,7 +24,7 @@ void HUD::Draw()
 			CAnimations* animations = CAnimations::GetInstance();
 			for (int i = 0; i < speedStack; i++)
 			{
-				float startX = 62 + i * 8;
+				float startX = 62 + (float)i * 8;
 				if (i == 6)
 					animations->Get(ID_ANI_SPEEDSTACK_MAX_P)->Render(Camera::GetInstance()->cam_x + startX + 4, Camera::GetInstance()->cam_y + 210);
 				else
